@@ -63,7 +63,7 @@
     "url": "https://github.com/elastic/kibana.git"
   },
   "dependencies": {
-    "@bigfunger/decompress-zip": "0.2.0-stripfix2",
+    "@bigfunger/decompress-zip": "0.2.0-stripfix3",
     "@spalger/angular-bootstrap": "0.12.1",
     "@spalger/filesaver": "1.1.2",
     "@spalger/leaflet-draw": "0.2.3",
@@ -77,7 +77,7 @@
     "angular-route": "1.4.7",
     "angular-sanitize": "1.5.7",
     "ansicolors": "0.3.2",
-    "autoprefixer": "5.1.1",
+    "autoprefixer": "6.3.7",
     "autoprefixer-loader": "2.0.0",
     "babel": "5.8.23",
     "babel-core": "5.8.23",
@@ -103,6 +103,7 @@
     "good": "6.3.0",
     "good-squeeze": "2.1.0",
     "gridster": "0.5.6",
+    "grunt-run": "0.6.0",
     "hapi": "8.8.1",
     "imports-loader": "0.6.4",
     "jade": "1.11.0",
@@ -114,8 +115,8 @@
     "json-stringify-safe": "5.0.1",
     "jstimezonedetect": "1.0.5",
     "leaflet": "0.7.5",
-    "less": "2.5.1",
-    "less-loader": "2.2.0",
+    "less": "2.7.1",
+    "less-loader": "2.2.3",
     "loader-utils": "0.2.11",
     "lodash": "3.10.1",
     "marked": "0.3.6",
@@ -143,28 +144,27 @@
     "angular-mocks": "1.4.7",
     "auto-release-sinon": "1.0.3",
     "babel-eslint": "4.1.7",
-    "chokidar": "1.0.5",
+    "chokidar": "1.6.0",
     "chromedriver": "2.22.1",
     "eslint": "1.5.1",
     "eslint-plugin-mocha": "1.0.0",
     "expect.js": "0.3.1",
     "faker": "1.1.0",
     "glob": "4.5.3",
-    "grunt": "0.4.5",
+    "grunt": "1.0.1",
+    "grunt-aws-s3": "0.14.5",
     "grunt-babel": "5.0.1",
     "grunt-cli": "0.1.13",
-    "grunt-contrib-clean": "0.6.0",
+    "grunt-contrib-clean": "1.0.0",
     "grunt-contrib-copy": "0.8.1",
     "grunt-esvm": "3.2.6",
-    "grunt-karma": "0.12.0",
-    "grunt-run": "0.5.0",
-    "grunt-s3": "0.2.0-alpha.3",
+    "grunt-karma": "0.12.2",
     "grunt-simple-mocha": "0.4.0",
     "gruntify-eslint": "1.0.1",
     "husky": "0.8.1",
     "intern": "3.2.3",
     "istanbul-instrumenter-loader": "0.1.3",
-    "karma": "0.13.22",
+    "karma": "1.2.0",
     "karma-chrome-launcher": "0.2.0",
     "karma-coverage": "0.5.1",
     "karma-firefox-launcher": "0.1.6",
@@ -177,14 +177,14 @@
     "makelogs": "2.0.0",
     "mocha": "2.3.4",
     "nock": "8.0.0",
-    "npm": "2.15.8",
+    "npm": "3.10.3",
     "portscanner": "1.0.0",
     "simple-git": "1.8.0",
     "sinon": "1.17.2",
     "source-map": "0.4.4"
   },
   "engines": {
-    "node": "4.4.7",
-    "npm": "2.15.8"
+    "node": "6.4.0",
+    "npm": "3.10.3"
   }
 }
